+++ packages/@webex/plugin-meetings/src/member/util.ts
@@ -16,6 +16,7 @@ import {
   _CALL_,
 } from '../constants';
 import ParameterError from '../common/errors/parameter';
+import {IMediaStatus} from './member.types';
 
 const MemberUtil: any = {};
 
@@ -252,6 +253,22 @@ MemberUtil.extractId = (participant: any) => {
   }
 
   return null;
+};
+
+/**
+ * extracts the media status from nested participant object
+ * @param {Object} participant the locus participant
+ * @returns {Object}
+ */
+MemberUtil.extractMediaStatus = (participant: any): IMediaStatus => {
+  if (!participant) {
+    throw new ParameterError('Media status could not be extracted, participant is undefined.');
+  }
+
+  return {
+    audio: participant.status.audioStatus,
+    video: participant.status.videoStatus,
+  };
 };
 
 /**