+++ packages/tools/package/test/integration/increment/increment.test.js
@@ -84,7 +84,7 @@ describe('increment', () => {
       patch: 3,
       release: 4,
       since: 'example-reference',
-      tag: 'example-tag',
+      tag: 'example/of/example-tag',
     };
 
     beforeEach(() => {
@@ -94,6 +94,7 @@ describe('increment', () => {
 
       spies.package = {
         inspect: spyOn(Package.prototype, 'inspect').and.callFake(function func() { return Promise.resolve(this); }),
+        syncVersion: spyOn(Package.prototype, 'syncVersion').and.callFake(function func() { return Promise.resolve(this); }),
         incrementVersion: spyOn(Package.prototype, 'incrementVersion').and.callFake(function func() { return this; }),
         apply: spyOn(Package.prototype, 'apply').and.callFake(function func() { return Promise.resolve(this); }),
       };
@@ -104,6 +105,9 @@ describe('increment', () => {
 
       spies.process = {
         cwd: spyOn(process, 'cwd').and.returnValue(rootDir),
+        stdout: {
+          write: spyOn(process.stdout, 'write').and.callFake(() => undefined),
+        },
       };
     });
 
@@ -122,6 +126,16 @@ describe('increment', () => {
           });
         });
     });
+
+    it('should call "package.inspect()" for each located package', () => increment.handler(options)
+      .then(() => {
+        expect(spies.package.inspect).toHaveBeenCalledTimes(listResolve.length);
+      }));
+
+    it('should call "package.syncVersion()" for each located package', () => increment.handler(options)
+      .then(() => {
+        expect(spies.package.syncVersion).toHaveBeenCalledTimes(listResolve.length);
+      }));
 
     it('should not increment any packages when the version details provided are undefined', () => increment.handler({})
       .then((results) => {
@@ -149,6 +163,13 @@ describe('increment', () => {
           expect(results[1].name).toBe(targetPackages[1]);
         });
     });
+
+    it('should write the list of packages updated and their corresponding new versions', () => increment.handler({ ...options })
+      .then(() => {
+        const generatedString = options.packages.map((pack) => `${pack} => 0.0.0-${options.tag.split('/').pop()}.0`).join('\n');
+
+        expect(spies.process.stdout.write).toHaveBeenCalledOnceWith(generatedString);
+      }));
 
     it('should return all packages when packages is not provided', () => increment.handler({ ...options, packages: undefined })
       .then((results) => {